requirements.txt
Flask==1.1.2
numpy==1.18.5
datasets==1.1.2
requests==2.23.0
torchvision==0.7.0
transformers==3.3.0
dill==0.3.3
torch==1.6.0
dataclasses==0.8
PyYAML==5.4.1
